refactor(web): update imports to use new datetime utility paths
