+++ packages/web/src/__tests__/utils/dates/date.label.test.ts
@@ -1,4 +1,4 @@
-import { getTimesLabel } from "@web/common/utils/dates/web.date.util";
+import { getTimesLabel } from "@web/common/utils/datetime/web.date.util";
 
 const meridians = (label: string) =>
   (label.match(/am/gi) || label.match(/pm/gi) || []).length;
