+++ packages/web/src/views/Calendar/components/Draft/sidebar/hooks/useSidebarActions.ts
@@ -24,7 +24,7 @@ import {
   computeCurrentEventDateRange,
   computeRelativeEventDateRange,
   getDatesByCategory,
-} from "@web/common/utils/dates/web.date.util";
+} from "@web/common/utils/datetime/web.date.util";
 import { assembleDefaultEvent } from "@web/common/utils/event.util";
 import { getX } from "@web/common/utils/grid.util";
 import {
