+++ packages/web/src/views/Calendar/components/Event/Grid/GridEvent/GridEvent.tsx
@@ -12,7 +12,7 @@ import {
   ZIndex,
 } from "@web/common/constants/web.constants";
 import { Schema_GridEvent } from "@web/common/types/web.event.types";
-import { getTimesLabel } from "@web/common/utils/dates/web.date.util";
+import { getTimesLabel } from "@web/common/utils/datetime/web.date.util";
 import { isOptimisticEvent } from "@web/common/utils/event.util";
 import { getLineClamp } from "@web/common/utils/grid.util";
 import { isRightClick } from "@web/common/utils/mouse/mouse.util";
