+++ packages/web/src/views/Calendar/components/Grid/Columns/TimesColumn/TimesColumn.tsx
@@ -3,7 +3,7 @@ import React, { useEffect, useMemo, useState } from "react";
 import {
   getColorsByHour,
   getHourLabels,
-} from "@web/common/utils/dates/web.date.util";
+} from "@web/common/utils/datetime/web.date.util";
 import { Text } from "@web/components/Text";
 import { StyledDayTimes, StyledTimesLabel } from "./styled";
